presence_app_v4/presence_app_v4/auth_service/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-limiter==3.5.0
psycopg2-binary==2.9.9
pyjwt==2.8.0
werkzeug==3.0.1
gunicorn==21.2.0
